refactors/cleaning up of feedback button
