+++ src/app/index.tsx
@@ -1,3 +1,4 @@
 export { Header } from './header/Header';
 export { Content } from './content/Content';
 export { SidebarContainer } from './sidebar/containers/SidebarContainer';
+export { Feedback } from './misc/Feedback';
